chainId removed, disclaimer added
The signing function used in this guide (signTransaction)does not support the chainId parameter, and thus its use is non-functional. In light of sharing only functional content, the chainId parameter was removed and a disclaimer was added for users that plan to store signed transactions near contentious hard fork events.